fix: stop repositioning the calendar entry suggestions
The little popover window for auto-complete and suggestions of new
calendar entries responds to isStandard() but it's not really a standard
window. It's a popup which isStandard() should reply false to.

We can see if the window isMaximizable() instead to determine what to
do. Another option was checking for a window title which it didn't have.