feat: change to packer for manging nvim packages
The primary purpose is to use a lua based tool. The previous vim-plug
tool worked flawlessly.

This now causes a problem with automatic installs. The headless command
throws an error. But the next run of vim will work just fine. I poked
around and couldn't find a good solution to this. But it's fine for now.

References:
https://github.com/wbthomason/packer.nvim/issues/502