feat: remove prettier formatting & ruby diagnostics from ALE
We had to set solargraph to stop auto-installing because it installed in
isolation and didn't know the projects ruby or rubocop versions. This
means each project we want rubocop diagnostics from will need to have
solargraph in its bundle file. This might not work out long term, we
will try and see. It might be worth the LSP features like go references,
rename, definition, etc.

We won't use solargraph rubocop to format since most of our projects use
prettier.

Next we install vim-prettier which is smart about finding the prettier
binary and it gives us a formatting command which we then setup to run
before buffer write for Ruby files.