fix: disabling of solargraph formatting setting
You can see that it was disabled by checking the results of this
command when an LSP server is attached to a buffer.

:lua print(vim.inspect(vim.lsp.get_active_clients()))

The results are cryptic but now we see the server_capabilities don't
include formatting since we turned it off on initialization.

And the resolved capabilities say our client and server both agree that
document_formatting is not supported.