fix: stop using ALE plugin
Here are the remaining uses:

standardrb - i don't use this in any projects right now, will figure it
out without ALE when it comes up. looks like someone made a solargraph
plugin for it though.

and prettier and most auto-formatting tools do these for me. I'm sure I
was using it in other languages and will miss it, but, I'll address that
when it comes up:

removing_trailing_space
trim_whitespace