feat: don't create Prettier autocmds and try not to run them
I forked the vim-prettier plugin to create this new option to not create
the autocmnds. They were running on every single BufWritePre for most of
the types of files I work in daily. But, I use eslint to run prettier on
most of those projects.

With this change I have full control of when prettier runs.

I first tried to use the new lua vim API's to clear autocmds for this
group. If the group does not exist the command errors.

In this spot in the init they were not loaded so it was giving an error.
I then tried to use packer to ensure the plugin was loaded. I got that
working but the autocmds were still not loaded. Then I tried to use
vim's exists() function to only run the clear if the autocmds are
loaded. All of this didn't work for me. That's when I forked the plugin.