fix: universal-ctags, installation is broken with snap
The problem I was running into was ludovicchabant/vim-gutentags giving
an error anytime it ran in vim. I thought it was that plugin but it
turned out to be something else. I put this in my ctags config:

--_echo=we-ran-your-ctags-config

And anytime ctags is ran and loads the config file it should echo that.
Well, it wasn't. Then I tried to manually use --options to set my config
file and it was saying permission denied on reading that file. But the
permissions were correct.

I ended up rebuilding from source just to see and it solved the problem.