fix: disable swap files in vim
When I was switching my config from vimscript to lua I skipped options
that I didn't understand. Well, this option is definitely one I
understand now. Swap files are a huge mess in Vim. When opening a file
multiple times it asks about the swap file and what I should do since
another instance has the same file open. Well, if I'm opening a file
twice all I want is for Vim to load the last saved version.

I believe the terminus plugin I use is what makes that auto-reloading
for files work well. Here is the blurb:

> Allows Vim to receive FocusGained and FocusLost events, even in the
terminal and inside tmux. This is in turn used to fire the :checktime
command, which, in conjunction with the 'autoread', allows Vim to
automatically pick up changes made by other processes when switching to
and from Vim.

References:
https://github.com/wincent/terminus