feat: add use_stripe_cassette snippets, simple and object
They are hidden from cmp completion by default because they also show
up outside of tests/ folder. I tried to use a function to set hidden to
false when inside of a test folder but it doesn't work. I think it's
because the code is compiled to the value when it's loaded. I think it
needs to take a function work right.

To expand it in insert mode:
use_stripe_cassette<C-k>