fix: correct regex to figure out the fixture type
This has been working to trigger fixture names:

people(:jimbo)
        people(:jimbo)

It regex captures "people" and uses that to open the people.yml fixture
file and look for the "jimbo:" entry.

But this doesn't work. It is capturing the fixture type of "person =
people".

person = people(:jimbo)

The new regex looks for end of string (: and captures everything before
it until the first space.