feat: manage stylua via lsp with null-ls
So, now that two lua formatters exists the command will popup a window
asking which server to use. Instead we just let it run them all to avoid
the popup. sumneko_lua won't do anything since it's disabled for
formatting.

Used :Mason to install stylua now that the stylua-nvim plugin doesn't do
it anymore.

We are dropping all the .editorconfig for lua because those were rules
for sumneko_lua formatting.