feat: add inlay hints if lsp server support it
Had to switch to neovim nightly to get inlay hints support.

Something in neovim nightly broke telescope file preview but it was
fixed in telescope main. Current projects don't have a new enough

typescript install and it throws an error but also fails in an unusable
way, so, let's disable it in tsserver for now.