feat(open): add git commit browser support to the open command
Introduce a new feature in the `open` command. If the argument provided
is a valid git commit hash, it will open the commit in the browser. If
not, it will open the argument in the default system `open` application.
This feature is implemented in a new bash script located in `bin/open`.

The `install.sh` script is also updated to create a symlink to the
new `open` command in `/usr/local/bin`, ensuring it's available in
the system path.