fix: stop including LSP percent for Ruby LSP
They already include the percent in the message so we see two
percentages during progress updates. If the LSP gives the percentage
they should update the client to stop including it in the message
string.

This is the default formatter function form the plugin with the LSP
check lines added.